refactor: рефактор страницы Catalog по замечаниям
