+++ src/api/catalog-article-page.js
@@ -5,7 +5,7 @@ function getCatalogArticlePageData({ articleId } = {}) {
   return axios
     .get(`${baseURL}${apiUrl}/catalog/${articleId}`)
     .then((response) => response.data)
-    .catch((err) => Promise.reject(new Error(`${err.message}`)));
+    .catch((err) => Promise.reject(err?.response));
 }
 
 export default getCatalogArticlePageData;
