+++ src/pages/Catalog/index.js
@@ -2,7 +2,7 @@ import BasePage from '../../layout/BasePage';
 import CardsSectionWithLines from '../../components/CardsSectionWithLines/CardsSectionWithLines';
 import AnimatedPageContainer from '../../components/AnimatedPageContainer/AnimatedPageContainer';
 import CardCatalog from '../../components/Cards/CardCatalog/CardCatalog';
-import { Loader, TitleH1, TitleH2 } from '../../components/utils/index';
+import { Loader, TitleH1, TitleH2 } from '../../components/utils';
 
 export {
   BasePage,
