+++ src/pages/CatalogArticle/locales/RU.js
@@ -1,6 +1,7 @@
 const catalogArticlePageTexts = {
   headTitle: 'Статья',
   headDescription: 'Полезная статья',
+  stubButtonText: 'Вернуться назад',
 };
 
 export default catalogArticlePageTexts;
